Update WealthLens AI

+++ member4/database.py
@@ -15,6 +15,8 @@ def create_database():
 
         id INTEGER PRIMARY KEY AUTOINCREMENT,
 
+        user_id INTEGER DEFAULT 0,
+
         monthly_investment REAL,
 
         annual_return REAL,
@@ -35,7 +37,7 @@ def create_database():
 # -----------------------------
 # Save User Data
 # -----------------------------
-def save_data(investment, rate, years, goal, future):
+def save_data(investment, rate, years, goal, future, user_id=0):
 
     conn = sqlite3.connect("wealth.db")
 
@@ -44,6 +46,7 @@ def save_data(investment, rate, years, goal, future):
     cursor.execute("""
     INSERT INTO wealth_data(
 
+        user_id,
         monthly_investment,
         annual_return,
         years,
@@ -52,9 +55,10 @@ def save_data(investment, rate, years, goal, future):
 
     )
 
-    VALUES(?,?,?,?,?)
+    VALUES(?,?,?,?,?,?)
 
     """, (
+        user_id,
         investment,
         rate,
         years,
@@ -67,15 +71,18 @@ def save_data(investment, rate, years, goal, future):
 
 
 # -----------------------------
-# Read All Data
+# Read User Data
 # -----------------------------
-def get_data():
+def get_data(user_id=None):
 
     conn = sqlite3.connect("wealth.db")
 
     cursor = conn.cursor()
 
-    cursor.execute("SELECT * FROM wealth_data")
+    if user_id is not None:
+        cursor.execute("SELECT * FROM wealth_data WHERE user_id=?", (user_id,))
+    else:
+        cursor.execute("SELECT * FROM wealth_data")
 
     data = cursor.fetchall()
 
@@ -85,15 +92,18 @@ def get_data():
 
 
 # -----------------------------
-# Delete All Data
+# Delete User Data
 # -----------------------------
-def delete_data():
+def delete_data(user_id=None):
 
     conn = sqlite3.connect("wealth.db")
 
     cursor = conn.cursor()
 
-    cursor.execute("DELETE FROM wealth_data")
+    if user_id is not None:
+        cursor.execute("DELETE FROM wealth_data WHERE user_id=?", (user_id,))
+    else:
+        cursor.execute("DELETE FROM wealth_data")
 
     conn.commit()
     conn.close()
@@ -102,4 +112,4 @@ def delete_data():
 # -----------------------------
 # Create Database Automatically
 # -----------------------------
-create_database()
+create_database()
